+++ packages/api-plugin-fulfillment/src/mutations/updateFulfillmentOptionsForGroup.js
@@ -77,7 +77,7 @@ export default async function updateFulfillmentOptionsForGroup(context, input) {
 
   const commonOrder = await context.queries.getCommonOrderForCartGroup(context, { cart, fulfillmentGroupId: fulfillmentGroup._id });
   // In the future we want to do this async and subscribe to the results
-  const rates = await context.queries.getFulfillmentMethodsWithQuotes(commonOrder, context);
+  const rates = await context.queries.fulfillmentMethodsWithQuotes(commonOrder, context);
 
   const { shipmentQuotes, shipmentQuotesQueryStatus } = getShipmentQuotesQueryStatus(rates);